# src/Client/Logistics.DriverApp/composeApp/src/openapi/api-spec.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Logistics.API",
    "version": "1.0"
  },
  "paths": {
...
    "/stats/daily-grosses": {
      "get": {
        "tags": [
          "Stat"
        ],
        "operationId": "GetDailyGrosses",
        "parameters": [
          {
            "name": "TruckId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "UserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "StartDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "EndDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DailyGrossesDtoResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyGrossesDtoResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyGrossesDtoResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/stats/monthly-grosses": {
      "get": {
        "tags": [
          "Stat"
        ],
        "operationId": "GetMonthlyGrosses",
        "parameters": [
          {
            "name": "TruckId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "UserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "StartDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "EndDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyGrossesDtoResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyGrossesDtoResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyGrossesDtoResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/stats/company": {
      "get": {
        "tags": [
          "Stat"
        ],
        "operationId": "GetCompanyStats",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/GetCompanyStatsQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyStatsDtoResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyStatsDtoResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyStatsDtoResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/stats/trucks": {
      "get": {
        "tags": [
          "Stat"
        ],
        "operationId": "GetTrucksStats",
        "parameters": [
          {
            "name": "StartDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "EndDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TruckStatsDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckStatsDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckStatsDtoPagedResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/stats/driver/{userId}": {
      "get": {
        "tags": [
          "Stat"
        ],
        "operationId": "GetDriverStats",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "UserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DriverStatsDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverStatsDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverStatsDtoPagedResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
...
    "/trucks": {
      "get": {
        "tags": [
          "Truck"
        ],
        "operationId": "GetTrucks",
        "parameters": [
          {
            "name": "IncludeLoads",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TruckDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckDtoPagedResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Truck"
        ],
        "operationId": "CreateTruck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTruckCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTruckCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTruckCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "required": [
          "city",
          "country",
          "line1",
          "state",
          "zipCode"
        ],
        "type": "object",
        "properties": {
          "line1": {
            "type": "string",
            "nullable": true
          },
          "line2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CompanyStatsDto": {
        "type": "object",
        "properties": {
          "ownerName": {
            "type": "string",
            "nullable": true
          },
          "employeesCount": {
            "type": "integer",
            "format": "int32"
          },
          "managersCount": {
            "type": "integer",
            "format": "int32"
          },
          "dispatchersCount": {
            "type": "integer",
            "format": "int32"
          },
          "driversCount": {
            "type": "integer",
            "format": "int32"
          },
          "trucksCount": {
            "type": "integer",
            "format": "int32"
          },
          "thisWeekGross": {
            "type": "number",
            "format": "double"
          },
          "thisWeekDistance": {
            "type": "number",
            "format": "double"
          },
          "lastWeekGross": {
            "type": "number",
            "format": "double"
          },
          "lastWeekDistance": {
            "type": "number",
            "format": "double"
          },
          "thisMonthGross": {
            "type": "number",
            "format": "double"
          },
          "thisMonthDistance": {
            "type": "number",
            "format": "double"
          },
          "lastMonthGross": {
            "type": "number",
            "format": "double"
          },
          "lastMonthDistance": {
            "type": "number",
            "format": "double"
          },
          "totalGross": {
            "type": "number",
            "format": "double"
          },
          "totalDistance": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "CompanyStatsDtoResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/CompanyStatsDto"
          }
        },
        "additionalProperties": false
      },
...
      "CreateTruckCommand": {
        "type": "object",
        "properties": {
          "truckNumber": {
            "type": "string",
            "nullable": true
          },
          "truckType": {
            "$ref": "#/components/schemas/TruckType"
          },
          "mainDriverId": {
            "type": "string",
            "format": "uuid"
          },
          "carHaulerMaxVehicles": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DailyGrossDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "driverShare": {
            "type": "number",
            "format": "double"
          },
          "gross": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "DailyGrossesDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyGrossDto"
            },
            "nullable": true
          },
          "totalGross": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "totalDistance": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DailyGrossesDtoResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/DailyGrossesDto"
          }
        },
        "additionalProperties": false
      },
...
      "DriverStatsDto": {
        "type": "object",
        "properties": {
          "thisWeekGross": {
            "type": "number",
            "format": "double"
          },
          "thisWeekShare": {
            "type": "number",
            "format": "double"
          },
          "thisWeekDistance": {
            "type": "number",
            "format": "double"
          },
          "lastWeekGross": {
            "type": "number",
            "format": "double"
          },
          "lastWeekShare": {
            "type": "number",
            "format": "double"
          },
          "lastWeekDistance": {
            "type": "number",
            "format": "double"
          },
          "thisMonthGross": {
            "type": "number",
            "format": "double"
          },
          "thisMonthShare": {
            "type": "number",
            "format": "double"
          },
          "thisMonthDistance": {
            "type": "number",
            "format": "double"
          },
          "lastMonthGross": {
            "type": "number",
            "format": "double"
          },
          "lastMonthShare": {
            "type": "number",
            "format": "double"
          },
          "lastMonthDistance": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "DriverStatsDtoPagedResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DriverStatsDto"
            },
            "nullable": true
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "EmployeeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "lastKnownLocation": {
            "type": "string",
            "nullable": true
          },
          "lastKnownLocationLat": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lastKnownLocationLng": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "salary": {
            "type": "number",
            "format": "double"
          },
          "salaryType": {
            "$ref": "#/components/schemas/SalaryType"
          },
          "joinedDate": {
            "type": "string",
            "format": "date-time"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GeoPoint": {
        "type": "object",
        "properties": {
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "GetCompanyStatsQuery": {
        "type": "object",
        "additionalProperties": false
      },
...
      "LoadDto": {
        "required": [
          "destinationAddress",
          "destinationLocation",
          "originAddress",
          "originLocation"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "originAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "originLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "destinationAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "destinationLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "deliveryCost": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "dispatchedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "pickedUpAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "canConfirmPickUp": {
            "type": "boolean"
          },
          "canConfirmDelivery": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/LoadStatus"
          },
          "type": {
            "$ref": "#/components/schemas/LoadType"
          },
          "assignedDispatcherId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "assignedDispatcherName": {
            "type": "string",
            "nullable": true
          },
          "assignedTruckId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "assignedTruckNumber": {
            "type": "string",
            "nullable": true
          },
          "currentAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "currentLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDto"
          },
          "tripId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "tripNumber": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "tripName": {
            "type": "string",
            "nullable": true
          },
          "assignedTruckDriversName": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MonthlyGrossDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "driverShare": {
            "type": "number",
            "format": "double"
          },
          "gross": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "MonthlyGrossesDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyGrossDto"
            },
            "nullable": true
          },
          "totalGross": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "totalDistance": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "MonthlyGrossesDtoResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/MonthlyGrossesDto"
          }
        },
        "additionalProperties": false
      },
...
      "Result": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "TruckDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "vehicleCapacity": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/TruckType"
          },
          "status": {
            "$ref": "#/components/schemas/TruckStatus"
          },
          "currentAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "currentLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "mainDriver": {
            "$ref": "#/components/schemas/EmployeeDto"
          },
          "secondaryDriver": {
            "$ref": "#/components/schemas/EmployeeDto"
          },
          "loads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoadDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TruckDtoPagedResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TruckDto"
            },
            "nullable": true
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "TruckStatsDto": {
        "type": "object",
        "properties": {
          "truckId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "truckNumber": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "gross": {
            "type": "number",
            "format": "double"
          },
          "driverShare": {
            "type": "number",
            "format": "double"
          },
          "drivers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeDto"
            },
            "nullable": true
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TruckStatsDtoPagedResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TruckStatsDto"
            },
            "nullable": true
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TruckStatus": {
        "enum": [
          "available",
          "en_route",
          "loading",
          "unloading",
          "maintenance",
          "out_of_service",
          "offline"
        ],
        "type": "string"
      },
      "TruckType": {
        "enum": [
          "flatbed",
          "freight_truck",
          "reefer",
          "tanker",
          "box_truck",
          "dump_truck",
          "tow_truck",
          "car_hauler"
        ],
        "type": "string"
      },
...
    }
  }
}